# specs/002-meeting-room-booking/system-analyze/_prompt-consistency/openapi.run-C.yaml
openapi: 3.0.3
info:
  title: 內部共享會議室預約系統 API
  version: 0.1.0-draft
  description: |
    功能分支：`002-meeting-room-booking`

    **狀態**：草稿

    ## API Schema 描述

    | 欄位 | 內容 |
    | --- | --- |
    | 共通 Header | JSON request 使用 `Content-Type: application/json`；登入後由瀏覽器自動帶入 HttpOnly session cookie；所有受保護 Endpoint 以該 session 辨識使用者 |
    | 時間格式 | 瞬間使用 ISO-8601 with offset（例如 `2026-07-31T14:00:00+08:00`）或 UTC `Z`；日期 query 使用 `YYYY-MM-DD`，所有「今日／平日／營業時段」以 `Asia/Taipei` 解讀 |

    ## 假設

    - 所有 `/api` Endpoint 除 `POST /api/auth/login` 外皆要求有效 session cookie；未登入固定回 `401 UNAUTHENTICATED`
    - 管理操作只允許 `facility_admin`；`employee` 與 `manager` 權限相同，越權固定回 `403 FORBIDDEN`
    - 時段採半開區間 `[starts_at, ends_at)`；相鄰但不重疊的預約或維護時段可成立
    - `POST /api/bookings` 的停用、容量、跨日、週末／假日、營業時段與時長違規使用 `400 BOOKING_RULE_VIOLATION`；設備需求欄位僅記錄；預約或維護時段重疊使用 409
    - 日曆與列表直接共用 `GET /api/bookings` 回應；第一版不另設 view 參數
    - `GET /api/bookings/mine` 第一版回傳全部歷史，不做分頁；預設依開始時間降冪排序
    - `GET /api/overview/today` 的 `busy_ratio` 以平日營業窗 09:00–21:00 共 720 分鐘計算；週末或假日的 `business_minutes` 與 `busy_ratio` 皆為 0
    - 維護時段查詢對所有已登入角色開放，讓預約畫面可先提示不可用；建立維護時段仍限設施管理員
...
paths:
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: 帳號密碼登入
      description: |
        驗證帳密、建立伺服端 session，並以 `Set-Cookie` 回傳 HttpOnly session cookie。

        **對應 FR**：US1-FR1（支援帳號密碼登入並區分三種角色）

        **設計備註**：帳號或密碼錯誤統一回 `INVALID_CREDENTIALS`，避免揭露帳號是否存在；成功後輪替 session ID。

        **測試規劃**
        - 有效帳密登入成功 → 200
        - 缺少帳號或密碼 → 400
        - 帳號或密碼錯誤 → 401
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: alice
              password: training-password
      responses:
        '200':
          description: 登入成功，回傳使用者資料並設定 session cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticatedUserResponse'
              example:
                authenticated: true
                user:
                  id: '11111111-1111-4111-8111-111111111111'
                  username: alice
                  display_name: Alice
                  role: employee
                  created_at: '2026-07-30T01:00:00.000Z'
        '400':
          description: 缺少帳號或密碼
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: VALIDATION_ERROR
                  message: username and password are required
                  details: []
        '401':
          description: 帳號或密碼錯誤
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: INVALID_CREDENTIALS
                  message: Invalid username or password
                  details: []
  /api/auth/logout:
    post:
      tags:
        - Auth
      summary: 登出
      description: |
        銷毀目前伺服端 session 並清除 cookie。

        **對應 FR**：US1-FR1（登入會話可結束並撤銷）

        **設計備註**：僅接受有效 session；成功回傳 JSON，不使用空 body。

        **測試規劃**
        - 已登入使用者登出成功 → 200
        - 未登入時要求登出 → 401
      operationId: logout
      responses:
        '200':
          description: 登出成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
              example:
                authenticated: false
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /api/auth/me:
    get:
      tags:
        - Auth
      summary: 取得目前登入使用者
      description: |
        取得目前 session 對應的使用者資料。

        **對應 FR**：US1-FR1（辨識登入使用者與角色）

        **設計備註**：不回傳 `password_hash`、session ID 或 cookie 值。

        **測試規劃**
        - 有效 session 取得本人資料 → 200
        - session 不存在或已失效 → 401
      operationId: getCurrentUser
      responses:
        '200':
          description: 成功取得目前使用者
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticatedUserResponse'
              example:
                authenticated: true
                user:
                  id: '11111111-1111-4111-8111-111111111111'
                  username: alice
                  display_name: Alice
                  role: employee
                  created_at: '2026-07-30T01:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthenticated'
...
  /api/bookings:
    post:
      tags:
        - Bookings
      summary: 建立預約
      description: |
        在交易內校驗並建立單次已確認預約。

        **對應 FR**
        - US1-FR3：必填用途、人數、起迄與設備需求
        - US1-FR4：規則通過後建立 `confirmed` 紀錄
        - US1-FR5：拒絕預約或維護時段重疊
        - US1-FR6：拒絕人數超過容量
        - GR-001：同室已確認預約不得重疊
        - GR-002：台北同一平日 09:00–21:00、30 分鐘至 4 小時且非假日
        - GR-003：停用會議室只拒絕新預約，不改動既有預約

        **設計備註**：`user_id` 取自 session，不接受 client 指定；時段採半開區間 `[starts_at, ends_at)`；預約衝突回 `409 BOOKING_CONFLICT`；維護重疊回 `409 MAINTENANCE_CONFLICT`；停用、假日、營業時段、時長、跨日或容量超限回 400；設備需求僅記錄不因「會議室缺該設備」而拒絕；DB exclusion 為併發最終防線。

        **測試規劃**
        - 合法且無衝突時建立已確認預約 → 201
        - 欄位缺漏、格式錯誤或人數超過容量 → 400
        - 停用會議室、跨日、假日、非平日或違反營業時段與時長 → 400
        - 未登入建立 → 401
        - room_id 不存在 → 404
        - 與已確認預約重疊（含併發競爭） → 409
        - 與維護時段重疊 → 409
      operationId: createBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
            example:
              room_id: '22222222-2222-4222-8222-222222222222'
              purpose: 產品週會
              attendee_count: 6
              needs_projector: true
              needs_video_conference: false
              starts_at: '2026-07-31T14:00:00+08:00'
              ends_at: '2026-07-31T15:00:00+08:00'
      responses:
        '201':
          description: 預約建立成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
              example:
                id: '33333333-3333-4333-8333-333333333333'
                room_id: '22222222-2222-4222-8222-222222222222'
                user_id: '11111111-1111-4111-8111-111111111111'
                purpose: 產品週會
                attendee_count: 6
                needs_projector: true
                needs_video_conference: false
                starts_at: '2026-07-31T14:00:00+08:00'
                ends_at: '2026-07-31T15:00:00+08:00'
                status: confirmed
                created_at: '2026-07-30T02:00:00.000Z'
                updated_at: '2026-07-30T02:00:00.000Z'
        '400':
          description: 欄位缺漏、格式錯誤、人數超過容量，或違反預約規則（停用、假日、營業時段、時長、跨日等）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: BOOKING_RULE_VIOLATION
                  message: Booking must be within 09:00–21:00 on one business day in Asia/Taipei
                  details:
                    - rule: GR-002
                      reason: outside_business_hours
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/RoomNotFound'
        '409':
          description: 與已確認預約或維護時段重疊
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                bookingConflict:
                  summary: 與已確認預約重疊
                  value:
                    error:
                      code: BOOKING_CONFLICT
                      message: The room already has a confirmed booking in this time range
                      details:
                        - room_id: '22222222-2222-4222-8222-222222222222'
                          starts_at: '2026-07-31T14:00:00+08:00'
                          ends_at: '2026-07-31T15:00:00+08:00'
                maintenanceConflict:
                  summary: 與維護時段重疊
                  value:
                    error:
                      code: MAINTENANCE_CONFLICT
                      message: The room is unavailable during a maintenance window
                      details:
                        - room_id: '22222222-2222-4222-8222-222222222222'
                          starts_at: '2026-07-31T14:00:00+08:00'
                          ends_at: '2026-07-31T15:00:00+08:00'

    get:
      tags:
        - Bookings
      summary: 依日期查詢已確認預約
      description: |
        依台北日曆日取得已確認預約，供日曆與列表共用。

        **對應 FR**
        - US2-FR2：提供日曆視圖資料
        - US2-FR3：提供列表視圖資料
        - US2-FR4：呈現已確認預約的會議室、時段及辨識摘要

        **設計備註**：`date` 必填並以 `Asia/Taipei` 日界查詢；`roomId` 可選；只回 `confirmed`，依 `starts_at`、會議室名稱排序；兩種視圖不得各自呼叫不同資料源。

        **測試規劃**
        - 查詢台北日全部已確認預約 → 200
        - 依 roomId 篩選 → 200
        - 該日無預約時回空陣列 → 200
        - date 缺漏或格式錯誤 → 400
        - 未登入查詢 → 401
        - roomId 不存在 → 404
      operationId: listBookingsByDate
      parameters:
        - name: date
          in: query
          required: true
          description: 台北日曆日，格式 `YYYY-MM-DD`
          schema:
            type: string
            format: date
            example: '2026-07-31'
        - name: roomId
          in: query
          required: false
          description: 可選的會議室 UUID 篩選
          schema:
            $ref: '#/components/schemas/Uuid'
          example: '22222222-2222-4222-8222-222222222222'
      responses:
        '200':
          description: 成功回傳指定日期的已確認預約
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingsByDateResponse'
              example:
                date: '2026-07-31'
                timezone: Asia/Taipei
                bookings:
                  - id: '33333333-3333-4333-8333-333333333333'
                    room_id: '22222222-2222-4222-8222-222222222222'
                    room:
                      name: 大會議A
                      floor: 3F
                    user_id: '11111111-1111-4111-8111-111111111111'
                    booked_by:
                      display_name: Alice
                    purpose: 產品週會
                    attendee_count: 6
                    needs_projector: true
                    needs_video_conference: false
                    starts_at: '2026-07-31T14:00:00+08:00'
                    ends_at: '2026-07-31T15:00:00+08:00'
                    status: confirmed
                    created_at: '2026-07-30T02:00:00.000Z'
                    updated_at: '2026-07-30T02:00:00.000Z'
        '400':
          description: date 缺漏或格式錯誤
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: VALIDATION_ERROR
                  message: date must use YYYY-MM-DD
                  details:
                    - field: date
                      reason: invalid_date
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/RoomNotFound'
  /api/bookings/mine:
    get:
      tags:
        - Bookings
      summary: 取得本人的預約歷史
      description: |
        取得目前登入使用者的預約歷史。

        **對應 FR**：US3-FR1（列出目前使用者自己的已確認與已取消預約）

        **設計備註**：不接受 `userId`；依 `starts_at` 降冪排序；`is_cancellable` 由狀態與目前時間衍生。

        **測試規劃**
        - 回傳本人的已確認與已取消預約 → 200
        - 本人無預約時回空陣列 → 200
        - 未登入查詢 → 401
      operationId: listMyBookings
      responses:
        '200':
          description: 成功回傳本人預約清單
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyBookingsResponse'
              example:
                bookings:
                  - id: '33333333-3333-4333-8333-333333333333'
                    room_id: '22222222-2222-4222-8222-222222222222'
                    room:
                      name: 大會議A
                      floor: 3F
                    user_id: '11111111-1111-4111-8111-111111111111'
                    purpose: 產品週會
                    attendee_count: 6
                    needs_projector: true
                    needs_video_conference: false
                    starts_at: '2026-07-31T14:00:00+08:00'
                    ends_at: '2026-07-31T15:00:00+08:00'
                    status: confirmed
                    is_cancellable: true
                    created_at: '2026-07-30T02:00:00.000Z'
                    updated_at: '2026-07-30T02:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /api/bookings/{bookingId}/cancel:
    post:
      tags:
        - Bookings
      summary: 取消預約
      description: |
        取消目前登入使用者擁有且尚可取消的預約。

        **對應 FR**
        - US3-FR2：可取消自己的未結束已確認預約
        - US3-FR3：取消後改為 `cancelled` 並釋放時段
        - US3-FR4：不得取消他人的預約

        **設計備註**：所有權不符回 403；已取消或已結束回 409；成功以交易更新狀態與 `updated_at`，保留歷史資料。

        **測試規劃**
        - 取消自己的未結束已確認預約 → 200
        - 未登入取消 → 401
        - 取消他人的預約 → 403
        - bookingId 不存在 → 404
        - 重複取消或取消已結束預約 → 409
      operationId: cancelBooking
      parameters:
        - $ref: '#/components/parameters/BookingIdPath'
      responses:
        '200':
          description: 預約取消成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
              example:
                id: '33333333-3333-4333-8333-333333333333'
                room_id: '22222222-2222-4222-8222-222222222222'
                user_id: '11111111-1111-4111-8111-111111111111'
                purpose: 產品週會
                attendee_count: 6
                needs_projector: true
                needs_video_conference: false
                starts_at: '2026-07-31T14:00:00+08:00'
                ends_at: '2026-07-31T15:00:00+08:00'
                status: cancelled
                created_at: '2026-07-30T02:00:00.000Z'
                updated_at: '2026-07-30T06:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          description: 不得取消他人的預約
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: FORBIDDEN
                  message: You cannot cancel another user's booking
                  details: []
        '404':
          $ref: '#/components/responses/BookingNotFound'
        '409':
          description: 預約已取消或已結束，無法再次取消
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: BOOKING_NOT_CANCELLABLE
                  message: Booking is already cancelled or has ended
                  details: []
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
      description: |
        HttpOnly session cookie，由 `connect-pg-simple` 管理。
        登入成功後瀏覽器自動帶入；所有受保護 Endpoint 以該 session 辨識使用者。
        僅 `POST /api/auth/login` 為公開端點，不需此 cookie。
...
  parameters:
...
    BookingIdPath:
      name: bookingId
      in: path
      required: true
      description: 預約 UUID
      schema:
        $ref: '#/components/schemas/Uuid'
      example: '33333333-3333-4333-8333-333333333333'
  responses:
    Unauthenticated:
      description: 未登入或 session 無效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHENTICATED
              message: Authentication is required
              details: []
...
    RoomNotFound:
      description: 會議室不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: ROOM_NOT_FOUND
              message: Room was not found
              details: []
    BookingNotFound:
      description: 預約不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: BOOKING_NOT_FOUND
              message: Booking was not found
              details: []
  schemas:
    Uuid:
      type: string
      format: uuid
      description: 對外資源 ID（UUID 字串）
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      description: |
        全檔共用的錯誤回應格式。
        各 Endpoint 仍逐一列出可能的 error status 與完整錯誤範例。
        `details` 可用於提供欄位或規則細節。
    ErrorBody:
      type: object
      required:
        - code
        - message
        - details
      properties:
        code:
          type: string
          description: 機器可讀的錯誤代碼
          enum:
            - VALIDATION_ERROR
            - INVALID_CREDENTIALS
            - UNAUTHENTICATED
            - FORBIDDEN
            - ROOM_NOT_FOUND
            - BOOKING_NOT_FOUND
            - BOOKING_RULE_VIOLATION
            - BOOKING_CONFLICT
            - MAINTENANCE_CONFLICT
            - BOOKING_NOT_CANCELLABLE
        message:
          type: string
          description: 人類可讀的錯誤訊息
        details:
          type: array
          description: 欄位、規則或衝突細節
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      description: |
        錯誤細節項目，依情境包含不同欄位：
        - 驗證錯誤：`field` + `reason`
        - 預約規則違反：`rule` + `reason`
        - 時段衝突：`room_id` + `starts_at` + `ends_at`
      oneOf:
        - $ref: '#/components/schemas/FieldErrorDetail'
        - $ref: '#/components/schemas/RuleErrorDetail'
        - $ref: '#/components/schemas/ConflictErrorDetail'
    FieldErrorDetail:
      type: object
      properties:
        field:
          type: string
          description: 驗證失敗的欄位名稱
        reason:
          type: string
          description: 驗證失敗原因
      required:
        - field
        - reason
      additionalProperties: false
    RuleErrorDetail:
      type: object
      properties:
        rule:
          type: string
          description: 違反的規則代碼（例如 GR-002）
        reason:
          type: string
          description: 規則違反原因
      required:
        - rule
        - reason
      additionalProperties: false
    ConflictErrorDetail:
      type: object
      properties:
        room_id:
          $ref: '#/components/schemas/Uuid'
        starts_at:
          type: string
          format: date-time
          description: 衝突時段開始瞬間
        ends_at:
          type: string
          format: date-time
          description: 衝突時段結束瞬間
      required:
        - room_id
        - starts_at
        - ends_at
      additionalProperties: false
    UserRole:
      type: string
      enum:
        - employee
        - manager
        - facility_admin
      description: 使用者角色
    User:
      type: object
      required:
        - id
        - username
        - display_name
        - role
        - created_at
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
          description: 登入使用者 UUID
        username:
          type: string
          description: 登入帳號
        display_name:
          type: string
          description: UI 顯示名稱
        role:
          $ref: '#/components/schemas/UserRole'
        created_at:
          type: string
          format: date-time
          description: 帳號建立時間
    AuthenticatedUserResponse:
      type: object
      required:
        - authenticated
        - user
      properties:
        authenticated:
          type: boolean
          description: 目前 request 是否具有效 session
        user:
          $ref: '#/components/schemas/User'
    LogoutResponse:
      type: object
      required:
        - authenticated
      properties:
        authenticated:
          type: boolean
          enum:
            - false
          description: 登出後為 false
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: 登入帳號
        password:
          type: string
          format: password
          description: 登入密碼
...
    BookingStatus:
      type: string
      enum:
        - confirmed
        - cancelled
      description: 預約狀態
    Booking:
      type: object
      required:
        - id
        - room_id
        - user_id
        - purpose
        - attendee_count
        - needs_projector
        - needs_video_conference
        - starts_at
        - ends_at
        - status
        - created_at
        - updated_at
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
          description: 預約 UUID
        room_id:
          $ref: '#/components/schemas/Uuid'
          description: 被預約的會議室 UUID
        user_id:
          $ref: '#/components/schemas/Uuid'
          description: 預約者 UUID，由 session 決定
        purpose:
          type: string
          description: 會議用途
        attendee_count:
          type: integer
          minimum: 1
          description: 預期人數
        needs_projector:
          type: boolean
          description: 是否需要投影機
        needs_video_conference:
          type: boolean
          description: 是否需要視訊設備
        starts_at:
          type: string
          format: date-time
          description: 預約開始瞬間
        ends_at:
          type: string
          format: date-time
          description: 預約結束瞬間
        status:
          $ref: '#/components/schemas/BookingStatus'
        created_at:
          type: string
          format: date-time
          description: 建立時間
        updated_at:
          type: string
          format: date-time
          description: 最後更新時間
    CreateBookingRequest:
      type: object
      required:
        - room_id
        - purpose
        - attendee_count
        - needs_projector
        - needs_video_conference
        - starts_at
        - ends_at
      properties:
        room_id:
          $ref: '#/components/schemas/Uuid'
          description: 被預約的會議室 UUID
        purpose:
          type: string
          description: 會議用途
        attendee_count:
          type: integer
          minimum: 1
          description: 預期人數
        needs_projector:
          type: boolean
          description: 是否需要投影機（僅記錄，不因會議室缺設備而拒絕）
        needs_video_conference:
          type: boolean
          description: 是否需要視訊設備（僅記錄，不因會議室缺設備而拒絕）
        starts_at:
          type: string
          format: date-time
          description: 預約開始瞬間
        ends_at:
          type: string
          format: date-time
          description: 預約結束瞬間
    RoomSummary:
      type: object
      required:
        - name
        - floor
      properties:
        name:
          type: string
          description: 會議室名稱
        floor:
          type: string
          description: 樓層標示
    BookedBySummary:
      type: object
      required:
        - display_name
      properties:
        display_name:
          type: string
          description: 預約者顯示名稱
    BookingWithRoomSummary:
      allOf:
        - $ref: '#/components/schemas/Booking'
        - type: object
          required:
            - room
            - booked_by
          properties:
            room:
              $ref: '#/components/schemas/RoomSummary'
            booked_by:
              $ref: '#/components/schemas/BookedBySummary'
    BookingsByDateResponse:
      type: object
      required:
        - date
        - timezone
        - bookings
      properties:
        date:
          type: string
          format: date
          description: 查詢的台北日曆日
        timezone:
          type: string
          example: Asia/Taipei
          description: 業務日曆時區
        bookings:
          type: array
          items:
            $ref: '#/components/schemas/BookingWithRoomSummary'
    MyBooking:
      allOf:
        - $ref: '#/components/schemas/Booking'
        - type: object
          required:
            - room
            - is_cancellable
          properties:
            room:
              $ref: '#/components/schemas/RoomSummary'
            is_cancellable:
              type: boolean
              description: 查詢衍生的可取消狀態，不落庫
    MyBookingsResponse:
      type: object
      required:
        - bookings
      properties:
        bookings:
          type: array
          items:
            $ref: '#/components/schemas/MyBooking'
...